Initial commit
Store-to-store shipping in bulk requires uploading an XLSX with
appropriate fields filled in.  This script takes 711's template,
fills it in, and saves to an output file for upload at:

https://myship2.7-11.com.tw/C2C/Import

We might want to automate the rest of the process in the future (i.e.
actually upload the file and get shipping stickers), but for now let's
just do the rest of the process manually.