Default to TSV, add short names to arguments
Since copying from Google Sheets results in a TSV string by default,
use tab as default separator.  But --separator can be used to change to
a comma if necessary.

Change --csv-fields to accept the argument specified multiple times
rather than require the user to provide a comma-separated string.
Rename to --field or -f.